+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 ## 0.6.0
+- Added Feedback API
+- Services that used to returns an array of Offer will now return OfferResult (ImageMatchingService, SearchService).
+- OfferResult now contains requestID and sessionID to be used with Feedback API and ImageMatchingService.
+- Fix a bug in deserialization of Offer when customIds contains a property with null as a value.  
 - Add the ability to use different endpoints.
 - Add HeaderMapper protocol to allow mapping of headers entries if the SDK is used with a proxy.
 - Remove unused code from URLBuilder.